requirements.txt
flask
flask-cors
flask-limiter
groq
python-dotenv
requests
ddgs
